15 sprint homework. 1. MediaPlayerRepository interface united
