--- app/src/main/java/com/example/playlistmaker/domain/repository/mediaplayer/PlayTrackRepository.kt
@@ -1,5 +1,0 @@
-package com.example.playlistmaker.domain.repository.mediaplayer
-
-interface PlayTrackRepository {
-    fun playTrack()
-}
