--- app/src/main/java/com/example/playlistmaker/domain/repository/mediaplayer/PreparePlayerRepository.kt
@@ -1,5 +1,0 @@
-package com.example.playlistmaker.domain.repository.mediaplayer
-
-interface PreparePlayerRepository {
-    fun preparePlayer(dataSource:String)
-}
